requirements.txt
fastapi
uvicorn
cryptography
pyjwt
pytest
pytest-cov
requests
httpx